# src/swagger/swagger.json
{
	"components": {
...
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "BOS config admin API routes",
		"version": "1.0.0",
		"description": "Welcome to the reference documentation for the Spinalcore Digital Twin REST API! </br></br>To learn about the common use cases and concept of Spinalcore REST APIs, check out our resource center  (<a>https://resourcecenter.fr.spinalcom.com/ressources-développeur</a>).</br></br>In addition to Spinalcore API Reference, we also provide complete documentation for BOS config adminnistration API, that need to install and implement Spinalcore Digital Twin Operating System on their built environment.",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Spinalcom",
			"email": "developers@spinalcom.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/add_digitaltwin": {
			"post": {
				"operationId": "AddDigitalTwin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"DigitalTwin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "set_as_actual_digitaltwin",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"url": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"url",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/get_all_digitaltwins": {
			"get": {
				"operationId": "GetAllDigitalTwins",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {},
											"type": "array"
										},
										{
											"properties": {
												"message": {}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"DigitalTwin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/get_digitaltwin/{digitaltwinId}": {
			"get": {
				"operationId": "GetDigitalTwin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"DigitalTwin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "digitaltwinId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/set_as_actual_digitaltwin/{digitaltwinId}": {
			"put": {
				"operationId": "SetActualDigitalTwin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"DigitalTwin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "digitaltwinId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/get_actual_digitaltwin": {
			"get": {
				"operationId": "GetActualDigitalTwin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"DigitalTwin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/get_digitaltwin_contexts": {
			"get": {
				"operationId": "GetDefaultDigitalTwinContexts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {},
											"type": "array"
										},
										{
											"properties": {
												"message": {}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"DigitalTwin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/get_digitaltwin_contexts/{digitaltwinId}": {
			"get": {
				"operationId": "GetDigitalTwinContexts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {},
											"type": "array"
										},
										{
											"properties": {
												"message": {}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"DigitalTwin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "digitaltwinId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/update_digitaltwin/{digitaltwinId}": {
			"put": {
				"operationId": "EditDigitalTwin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"DigitalTwin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "digitaltwinId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"url": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/delete_digitaltwin/{digitaltwinId}": {
			"delete": {
				"operationId": "RemoveDigitalTwin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"DigitalTwin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "digitaltwinId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/delete_actual_digitaltwin": {
			"delete": {
				"operationId": "RemoveActualDigitaTwin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"DigitalTwin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}